A bug-fix and feature bonanza!
So much stuff, can't remember it all...
* Added 'Last connected' info feature to Quick Info area
* Made Quick Note editing widget bigger, multi-line and with wrapping (NOW WITH BLINKING CURSOR!! WOO!!! :-P)
* Re-did tooltip code to eliminate quite a few unneeded lines of code
* Disabled and deprecated 'SSH password' entry / editing because most Unix-like systems don't include sshpass by default.  Just use private keys after setting up remote systems through an interactive SSH session
* Make timed scanning code a little more robust
* Bumped version to 0.5.0
* Replaced SV_ constants with FL_ constants where possible
* Window / taskbar icon is back, this time using standardized FLTK methods and not X11 hackery -- still doesn't look very good, though
* Code clean-up, more commenting and minor fixes

+++ src/vnc.h
@@ -86,13 +86,13 @@ public:
   rfbClient * vncClient;
   HostItem * itm;
   bool allowDrawing;
-  int waitTime;
+  uint16_t waitTime;
   int nLastClientWidth;
   int nLastClientHeight;
   Fl_RGB_Image * imgCursor;
   int nCursorXHot;
   int nCursorYHot;
-  int inactiveSeconds;
+  uint16_t inactiveSeconds;
   int nLastScrollX;
   int nLastScrollY;
   int centeredX;
